courses all done

+++ app/Http/Controllers/CourseController.php
@@ -68,6 +68,10 @@ class CourseController extends Controller
     public function store(Request $request)
     {
         $user = Auth::user();
+$user->is_admin=TRUE; //to be deleted in production        
+        if (!$user->is_admin){
+            return response()->json(['message'=>'Only administrators can create a new courses', 'code'=>403],403);
+        }
         $values = $request->all();
         $values['user_id'] = $user->id;
 
@@ -91,17 +95,17 @@ class CourseController extends Controller
      */
     public function show(Course $course)
     {
-//         $course = Course::with(['tracks'=> function ($query){  
-  //          $query -> with('unit')
-    //            ->select('description','id','track','level_id')->with('level')
-      //          ->with(['skills' => function ($query) {
-        //          $query->select('track_id','skill')->orderBy('skill_order');}])
-          //      ->orderBy('track_order'); 
-            //}])->find($id);
+         $course = Course::with(['tracks'=> function ($query){  
+            $query -> with('unit')
+                ->select('description','id','track','level_id')->with('level')
+                ->with(['skills' => function ($query) {
+                  $query->select('track_id','skill')->orderBy('skill_order');}])
+                ->orderBy('track_order'); 
+            }])->find($course->id);
 
-//        if (!$course) {
-  //          return response()->json(['message' => 'This course does not exist', 'code'=>404], 404);
-    //    }
+        if (!$course) {
+            return response()->json(['message' => 'This course does not exist', 'code'=>404], 404);
+        }
         return response()->json(['course'=>$course, 'code'=>201], 201);
     }
 
@@ -114,16 +118,15 @@ class CourseController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Course $course)
-    {
+    {        
         $logon_user = Auth::user();
 $logon_user->is_admin = TRUE; //to be deleted for live, this makes everyone admin
-return $request;
         if ($logon_user->id != $course->user_id && !$logon_user->is_admin) {            
             return response()->json(['message' => 'You have no access rights to update course','code'=>401], 401);     
         }
 
         if ($request->hasFile('image')) {
-            unlink('images/courses/'.$course->id.'.png'); 
+            if (file_exists('images/courses/'.$course->id.'png')) unlink('images/courses/'.$course->id.'.png'); 
             $file = $request->image->move(public_path('images/courses'), $course->id.'.png');
         } 
         
@@ -140,7 +143,8 @@ return $request;
      */
     public function updateImage(Request $request, Course $course)
     {
-return $course;        $logon_user = Auth::user();
+return $course;
+        $logon_user = Auth::user();
         if ($logon_user->id != $course->user_id && !$logon_user->is_admin) {            
             return response()->json(['message' => 'You have no access rights to update course image','code'=>401], 401);     
         }
@@ -161,11 +165,11 @@ return $course;        $logon_user = Auth::user();
     public function destroy(Course $course)
     {
         $logon_user = Auth::user();
-        if ($logon_user->id != $course->created_by()->user_id && !$logon_user->is_admin) {            
+        if ($logon_user->id != $course->user_id && !$logon_user->is_admin) {            
             return response()->json(['message' => 'You have no access rights to delete course','code'=>401], 401);
         } 
         if (sizeof($course->houses)>0){
-            return response()->json(['message'=>'There are classes based on this course. Delete those classes first.','code'=>409],409);
+            return response()->json(['message'=>'There are classes based on this course. Delete those classes first.','code'=>500],500);
         }
         $course->delete();
         return response()->json(['message'=>'Course '.$course->name.' deleted','code'=>201], 201);
